DATE type support
Add the DATE keyword. Modify the INSERT statement to allow date data that meets the conditions to be inserted into the table, and illegal dates will be rejected. Modify the WHERE clause to allow comparison of date type columns, and illegal dates will be rejected. Modify the B+ tree implementation to support date type comparison.

Signed-off-by: HanZiyao <h56983577@126.com>